+++ app/controllers/notifications/notifications_controller.rb
@@ -3,10 +3,7 @@ module Notifications
     def index
       @notifications = notifications.includes(:actor).order('id desc').page(params[:page])
 
-      unread_ids = []
-      @notifications.each do |n|
-        unread_ids << n.id unless n.read?
-      end
+      unread_ids = @notifications.reject(&:read?).select(&:id)
       Notification.read!(unread_ids)
 
       @notification_groups = @notifications.group_by { |note| note.created_at.to_date }
@@ -20,6 +17,7 @@ module Notifications
     private
 
     def notifications
+      raise "You need reqiure user login for /notifications page." unless current_user
       Notification.where(user_id: current_user.id)
     end
   end