+++ app/models/notification.rb
@@ -2,7 +2,5 @@
 class Notification < ActiveRecord::Base
   include Notifications::Model
 
-  self.per_page = 20
-
   # Write your custom methods...
 end