+++ lib/notifications.rb
@@ -2,15 +2,13 @@ require 'notifications/model'
 require 'notifications/engine'
 require 'notifications/configuration'
 require 'notifications/version'
-require 'will_paginate'
-require 'will_paginate/active_record'
+require 'kaminari'
 
 module Notifications
   class << self
     def config
       return @config if defined?(@config)
       @config = Configuration.new
-      @config.per_page                 = 32
       @config.user_class               = 'User'
       @config.user_name_method         = 'name'
       @config.user_avatar_url_method   = nil